Refactor typelist settings
